objectives/requirements.txt
Django==3.1.4
djangorestframework==3.12.2
pycodestyle==2.6.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
django-debug-toolbar==3.2
